Add - router within app refactored
