+++ src/components/Home.js
@@ -4,13 +4,16 @@ import banners from "../data/featured-banners.json";
 import products from "../data/featured-products.json";
 import FeaturedProducts from "./FeaturedProducts";
 import Categories from "./Categories";
+import Header from "../Header.js";
+import Footer from "../Footer.js";
+import Button from "./Button";
 
 const ImageInSlider = styled.img`
   max-height: 100vh;
   width: 100vh;
 `;
 
-function Home() {
+function Home({ showProducts, setShowProducts }) {
   const [index, setIndex] = React.useState(0);
 
   const Image = ({ src }) => {
@@ -29,16 +32,21 @@ function Home() {
   };
 
   return (
-    <div className="wrapper">
-      <div className="slider">
-        <button onClick={() => setIndex(index - 1)}>{"<"}</button>
-        <Slider images={images}>{/* <Images /> */}</Slider>
+    <div className="App">
+      <Header showProducts={showProducts} setShowProducts={setShowProducts} />
+      <Button setShowProducts={setShowProducts} showProducts={showProducts} />
+      <div className="wrapper">
+        <div className="slider">
+          <button onClick={() => setIndex(index - 1)}>{"<"}</button>
+          <Slider images={images}>{/* <Images /> */}</Slider>
 
-        <button onClick={() => setIndex(index + 1)}>{">"}</button>
+          <button onClick={() => setIndex(index + 1)}>{">"}</button>
+        </div>
+
+        <Categories />
+        <FeaturedProducts featuredProducts={products.results} />
       </div>
-
-      <Categories />
-      <FeaturedProducts featuredProducts={products.results}/>
+      <Footer />
     </div>
   );
 }
